implementando autorizacao para execucao dos metodos

+++ src/main/java/dive/tech/exercicios_mvc/controller/JogoController.java
@@ -1,6 +1,5 @@
 package dive.tech.exercicios_mvc.controller;
 
-import dive.tech.exercicios_mvc.model.entity.Genero;
 import dive.tech.exercicios_mvc.model.entity.Jogo;
 import dive.tech.exercicios_mvc.model.service.JogoService;
 
@@ -8,7 +7,6 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.validation.Valid;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Response;
-import java.util.List;
 
 @ApplicationScoped
 @Path("/jogo")
@@ -33,7 +31,15 @@ public class JogoController {
     @POST
     @Consumes("application/json")
     @Produces("application/json")
-    public Response createJogo(@Valid Jogo jogo) {
+    public Response createJogo(@Valid Jogo jogo,
+                               @HeaderParam("Authorization") String authorization) {
+
+        if(!"Bearer codigo123".equalsIgnoreCase(authorization)) {
+            return Response
+                    .status(Response.Status.NOT_FOUND)
+                    .build();
+        }
+
         try {
             Jogo jogoCriado = jogoService.createJogo(jogo);
             return Response
@@ -52,7 +58,15 @@ public class JogoController {
     @Consumes("application/json")
     @Produces("application/json")
     public Response updateJogo( @PathParam("id") Long id,
+                                @HeaderParam("Authorization") String authorization,
                                 String nome) {
+
+        if(!"Bearer codigo123".equalsIgnoreCase(authorization)) {
+            return Response
+                    .status(Response.Status.NOT_FOUND)
+                    .build();
+        }
+
         Jogo jogoAtualizado = jogoService.updateJogo(nome,id);
 
         try {
@@ -70,7 +84,15 @@ public class JogoController {
     @DELETE
     @Path("{id}")
     @Produces("application/json")
-    public Response deleteJogo(@PathParam("id") Long id) {
+    public Response deleteJogo(@PathParam("id") Long id,
+                               @HeaderParam("Authorization") String authorization) {
+
+        if(!"Bearer codigo123".equalsIgnoreCase(authorization)) {
+            return Response
+                    .status(Response.Status.NOT_FOUND)
+                    .build();
+        }
+
         Jogo jogoDeletado = jogoService.deleteJogo(id);
 
         try {
